+++ CHANGELOG.md
@@ -1,7 +1,16 @@
 # Changelog
 
+## master
+
+### Enhancements
+* Added `secret` subcommand
+
+### Breaking
+* Changed structure of the theme
+
 ## 0.1.0
 
 ### Enhancements
 * Added `confirm` subcommand
+* Added `input` subcommand
 * Added a theme for `dialoguer` library